Introduce a helper function get_bit
Actually, the low-level compression scheme used here is
context-adaptive binary arithmetic coding (known as CABAC, also used in H.264).
The idea is to combine the effectiveness of arithmetic coding with simple context
adaption.

A standard binary coder without context adaption has a fixed table with
probabilities for the different symbols to decode, and can work fairly
efficient, for example by complete lookup tables or binary search in a
sorted array of accumulated probabilities. Updating those tables is an
expensive operation, so you can not easily update the probabilities
after every decoded symbol.

In the special case that there are only two symbols to decode (that's what
the binary in "binary arithmetic coding" means), arithmetic coding is a
simple comparison against a single threshold value, and updating that
threshold value is as simple as it gets: You just update one number and
that's it.

In this decoder, there are a lot of different thresholds (actually around
8000 of them), all initialized with 0x400, which actually means 0.5, as the
context is rescaled by "range / 0x800". The threshold is the expected
probability of finding a zero, so it is raised if a zero is found and
lowered if a one is found. The algorithm is actually an exponentially
weighted moving average, calculating the new expected probability by
  new_probability = 31/32*old_probability + 1/32*current_sample
where current_sample is 1.0 if the current decoded bit is a zero, and
0.0 if the current decoded bit is a one. This "simplifies" to
  new_probability = 31/32*old_probability + 1/32   # found 0
or
  new_probability = 31/32*old_probability + 0/32   # found 1
You actually see that in the code. The calculation of 31/32*X
is performed as "X - X/32" with is "X - (X >> 5)". Note that
"X - ((X + 31) >> 5)" is the same, but rounding up instead of down.

This commit introduces the helper function and a use at a single place
to show how it is meant. Further uses of the function are introduced in the
next commit.