Add comments about liveness of ecx
I am going to transform ecx (divide its content by two) inside a certain
region.  The comments I add in this commit show the range of the region I
need to change.

A variable (or register) is said to be "dead" if the value of it at that
point is not going to be used in the following code.  This means that a
change of the way a variable is used before a point where this variable is
considered dead can not have any influence on any code after that point.
Also, anything stored in a variable before a point where that variable is
dead need not be considered if the use of the variable is changed after that
point.  So I am free to change the use of a variable between two points
where it is "dead" without having to worry about consequences outside the
area between those two points.

Note that declaring a variable dead needs to account any possible code flow.
The code I am working on has the nice property that all register-named
variables actually are dead at the beginning/end of the main loop, so they
are also dead at any "continue" statement.  The region I am going to change
can be entered from the "if" and the "else" branch of a conditional
statement.  The register is live (the opposite of dead, so it's value is
going to be used) at the end of the "if" branch as well as the end of the
"else" branch, so I have to look into both of those branches.