Rescale ecx by two in a certain region
All accesses to ecx in the region marked by the comments introduced in the
previous commit have been changed. As explained in the previous commit,
these changes together are stand-alone and don't affect the remaining code.

This commit also removes the range-delimiting comments again, as their job
is done now.