Rescale state.sp_30 in the top "if" branch.
This one is actually interesting. sp_30 is obviously live at the point
marked 0046808F, as its value is used in the else branch of the if clause at
this point.  The block after the point marked 0046804A is *conditionally*
executed, so the assignment to sp_30 in that block might be skipped, and it
seems that sp_30 is live at 0046804A.

Actually, this is not the case, because the condition at 0046808F can not
be true for the first iteration of the loop (sp_18 still has its initial
value of 0). So in the first iteration of the main loop, sp_30 is not live
at 0046808F, as the if branch is taken. In all further iterations, sp_30
might be live, but in those iterations, out_pos is not zero, so the
conditional block at 0046804A is executed, overwriting sp_30. Beacuse of
this, sp_30 is not live at 0046804A.