Introduce helpers for get_bit loops.
These multi-number getters use a whole set of contexts, starting at a base offset
given as parameter:
 - the first bit is fetched with the context at offset 1
 - the second bit is fetched with the context at offset 2 or 3
 - the third bit is fetched with the context at offset 4,5,6 or 7
So all the bits already fetched determine the context to use for the next bit.
This is equivalent to maintaining probabilities for all possible results and
doing an n-bit arithmetic lookup, but has the advantage of easy just-in-time
updates.

Interestingly, some of the numbers are stored MSB first and other numbers are
stored LSB first. Probably this is to optimize convergence of some heavily
skewed probabilities. Take for example this distribution:

0: 40% 1: 40% 2: 10% 3: 10%

If you decode it MSB first, the decoder converges to

0/1: 80% -> 0: 50%, 1: 50%
2/3: 20% -> 2: 50%, 3: 50%

Only one context needs to be adapted (0 vs. 1 and 2 vs. 3 stay at the initial
50% ratio). Each decoded number helps pushing the first context towards 80% 0/1.

On the other hand, if you decode it LSB first, the decoder converges to

0/2: 50% -> 0: 80%, 2: 20%
1/3: 50% -> 1: 80%, 3: 20%

Each sample either contributes to the 0 vs. 2 or the 1 vs. 3 context. So the
rate of convergence is only half of the rate in the MSB decoding scenario.

So it makes sense to start decoding the bits with least entropy first, to help
their context(s) to converge quickly.