Refactor code to get a byte, while looking at a "reference" byte.
The idea of that code is that the byte to fetch is quite likely similar to
a byte already fetched, likely at the distance of the last copy operation.
To exploit the similarities, there are different contexts for fetching a bit
for the new number depending on whether the corresponding bit of the
reference byte is 0 or 1. (The set of contexts if the reference bit is
zero is at offset 0x100 compared to some base, while an offset of 0x200
is used if the reference bit is one). It is expected that the probabilities
are strongly skewed towards the new bit being equal to the corresponding
bit of the reference byte.

As soon as a mismatch is found, the decoder switches to a third set of
contexts, at offset 0 relative to the base value. This is done because
as soon as a bit in the new byte does not match the corresponding bit
in the old byte, it is expected that the less significant bits also
likely don't match, and decoding them using the same contexts would reduce
the skew of the contexts. (Remember: A big skew (most bits decoded via
using a specific context have the same value) is needed to get high
efficiency from arithmetic coding)