Split dual-purpose use of sp_18
state.sp_18 is used to carry the index of a "decoder state" that is used to
choose between different contexts (It actually is a number between 0 and 11,
so there are just 12 states). This state is persistent over iterations of
the main loop.

But sp_18 is also used to carry a flag meaning (identified as
"fetch_new_distance", but the transformation in this commit would work
equally well if the purpose of this flag is not yet known). To set this
flag, 12 (0xC) is added to sp_18, so sp_18 is now in the disjunct range
12 to 23. Before the end of the loop, sp_18 is either reinitialized with
a constant in the original state range (actually 7 or 10), or the flag
is "removed" by subtracting 0xC again.