get rid of sp_18 and sp_30
sp_18 is (as made even more clear in the previous two commits) the number of
some state of the decoder, while sp_30 is used for two different purposes
inside a single operation of the loop.  As sp_30 is dead at the end of an
iteration, it does not make sense to keep it around in the state structure,
a local variable does do the job perfectly. Because sp_30 serves two
different purposes, use two different local variables.