eliminate a common subexpression for position-dependent contexts
at two points, the context used for arithmetic decoding depends not only on
the internal state, but also in the position in a 32-bit word. This essentially
creates four different contexts for the different bytes of a 32-bit-word, which
makes a lot of sense if the data to be compressed is composed of 32-bit entitites.

Interestingly, state is shifted 4 bits (two would be enough), so there would be
room for 16 different contexts dependent on the position inside a 16-byte sequence,
but the later 12 contexts are not used at all. The range of refined_state_nr is
0x00 .. 0xB3, but the code reserves all the way up to 0xBF in the context array.