Replace access to yet-uninitialized elements by returning a constant
Previously, the code used the fact that untouched elements in a byte array
are zero.  Effectively, the decoded file is preceeded by a virtual run of
zeroes as long as the decoded file itself (the assumption, that was already
used in some transforms, is that the distance never exceeds the file
length).  Having a lot of zeroes available as "previously decoded bytes"
allows compact encoding of zero runs near the begin of the file.  This is
why these virtual zeroes before the beginning are very common in LZ77
coding variants.