Introduce state class
This commit introduces a State class, and has the decoder create 12 instances
of it, which are properly assigned to variables or linked to get the state
machine behaviour obtained from state_nr before.

With the state class containing the state-dependent contexts of the decoder,
the "thresholds" array can be removed from the ArithDecoder class.